Set up Shellcheck as a dedicated backend (#7)
We'll soon be adding Shellfmt as an example formatter. For formatters, you must set up some boilerplate for your language if it doesn't yet have it, such as defining `BashFormatterTargets`. We don't want that to live in the same file as `shellfmt.py` because it would be too noisy. So, we're going to want a dedicated folder for shellfmt, which means shellcheck should be similar.

Given that we want to eventually publish this example repo, this is also useful behavior to make users opt-in to Shellcheck, rather than being automatically activated.