Add a naive `binary` implementation (#9)
Corresponds to https://www.pantsbuild.org/v2.0/docs/plugins-run-goal.

Unlike Pex, running `dist/<new_bash_binary>` will not actually run the script, as there is no equivalent to Pex for Bash that we know of. Instead, we simply create a Zip file with all the relevant sources included. The onus is then on the user to unzip the file and run whatever script they want. 

While not as useful as Pex, this implementation still is a solid example for how `binary` works. It's easier to follow than our Pex implementation, which is riddled with our Python abstractions like `TwoStepPexRequest`.