Add a `run` implementation for Bash (#10)
Corresponds to https://www.pantsbuild.org/v2.0/docs/plugins-run-goal (still WIP).

Now Pants knows how to run arbitrary Bash scripts. Note that it will ignore the shebang and always run the equivalent `/path/to/bash script.sh`.

This implementation is extremely simplistic, as there isn't much to running a Bash script.

Originally, this implementation was going to add a `--bash-setup-executable-search-paths` option, but it looks like that is not necessary for `InteractiveProcess`; the `env` is not hermetic.