Fix and improve error messages for `zip` and `bash` not being found
Originally I was going to factor this code up in Pants, but I realized that that would result in worse error messages. We should keep the error messages local so that we can give the specific remedy for fixing the issue, which varies depending on how the plugin is implemented.